Move methods to util
