+++ app/client/events/guild-member-add.js
@@ -1,5 +1,6 @@
 'use strict'
 const { MessageEmbed } = require('discord.js')
+const { getOrdinalNum } = require('../../util').util
 
 const guildMemberAddHandler = async (client, member) => {
   if (member.user.bot) {
@@ -21,20 +22,6 @@ const guildMemberAddHandler = async (client, member) => {
   if (persistentRoles.size > 0) {
     member.roles.add(persistentRoles)
   }
-}
-
-function getOrdinalNum (number) {
-  let selector
-
-  if (number < 0) {
-    selector = 4
-  } else if ((number > 3 && number < 21) || number % 10 > 3) {
-    selector = 0
-  } else {
-    selector = number % 10
-  }
-
-  return number + ['th', 'st', 'nd', 'rd', ''][selector]
 }
 
 module.exports = guildMemberAddHandler
